+++ packages/mailflow-bundled/src/mailflow_bundled/__init__.py
@@ -22,6 +22,7 @@ from mailflow_mail_imap.plugin import plugin as imap_plugin
 from mailflow_notify_console.plugin import plugin as notify_plugin
 from mailflow_notify_onebot.plugin import plugin as onebot_plugin
 from mailflow_notify_openclaw_weixin.plugin import plugin as openclaw_weixin_plugin
+from mailflow_notify_openwechat.plugin import plugin as openwechat_plugin
 from mailflow_notify_wechaty.plugin import plugin as wechaty_plugin
 from mailflow_storage_sqlite.plugin import plugin as storage_plugin
 
@@ -36,6 +37,7 @@ BUNDLED_PLUGINS: tuple[Any, ...] = (
     notify_plugin,
     onebot_plugin,
     wechaty_plugin,
+    openwechat_plugin,
     openclaw_weixin_plugin,
     nonebot_export_plugin,
     astrbot_export_plugin,