+++ plugins/mailflow-notify-openwechat/src/mailflow_notify_openwechat/__init__.py
@@ -1,0 +1,5 @@
+"""openwechat WeChat gateway plugin."""
+
+from .plugin import OpenWechatPlugin, plugin
+
+__all__ = ["OpenWechatPlugin", "plugin"]